Agent as tool nd function tool/pyproject.toml
[project]
name = "agent-as-tool-nd-function-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.2",
    "python-dotenv>=1.1.1",
    "rich>=14.0.0",
]
